docs - updated documentation of queue in code comments

+++ data_structures/v2/include/queue.h
@@ -6,18 +6,24 @@
 
 #include "arena.h"
 
+// Node of the queue, stores the value and a reference to the next node
 typedef struct QueueNode {
     int value;
     struct QueueNode* next;
 } QueueNode;
 
+// Queue struct, stores references to the front and rear.
+// An array for the nodes of the queue (of size capacity on initialization) - node_pool
+// An array of free nodes (point to the free spaces in node_pool)
+// Size of enqueued elements
+// Total capacity
 typedef struct Queue {
     QueueNode* front;
     QueueNode* rear;
     QueueNode* node_pool;
     QueueNode* free_list;
-    size_t size;
-    size_t capacity;
+    int size;
+    int capacity;
 } Queue;
 
 bool queue_init(Queue* queue, mem_arena* arena, int capacity);
